Change checksum to a long
git-svn-id: svn://svn.freehep.org/svn/freehep/trunk/hepio/freehep-rootio@10726 52ca69fb-99d5-0310-9ee9-f1395e6d145f

+++ src/main/java/hep/io/root/daemon/xrootd/XrootdURLConnection.java
@@ -131,14 +131,14 @@ public class XrootdURLConnection extends URLConnection
       return getLastModified();
    }   
    
-   public int getCheckSum() throws IOException
+   public long getCheckSum() throws IOException
    {
       if (session == null) return -1;
       else
       {
          String result = session.query(XrootdProtocol.kXR_Qcksum,url.getFile());
          String[] split = result.split(" ");
-         return Integer.parseInt(split[1]);
+         return Long.parseLong(split[1]);
       }
    }
